Update Android scanner workflow IDs and add summaries
