+++ models/workflowbuilder.go
@@ -5,6 +5,7 @@ import bitriseModels "github.com/bitrise-io/bitrise/models"
 type workflowBuilderModel struct {
 	Steps       []bitriseModels.StepListItemModel
 	Description string
+	Summary     string
 }
 
 func newDefaultWorkflowBuilder() *workflowBuilderModel {
@@ -21,5 +22,6 @@ func (builder *workflowBuilderModel) generate() bitriseModels.WorkflowModel {
 	return bitriseModels.WorkflowModel{
 		Steps:       builder.Steps,
 		Description: builder.Description,
+		Summary:     builder.Summary,
 	}
 }
